requirements.txt
Werkzeug==2.2.2
flask==2.2.3
jsonschema==4.21.1
pyyaml==6.0.1
pytest==7.4.4